Refactor into API project
